refactor(probe): extract shared planning primitives module
